backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0


SQLAlchemy==2.0.13
psycopg2-binary==2.9.6
python-dotenv==1.0.0
alembic==1.11.1

pydantic[email]

passlib[bcrypt]
pytest==7.4.0
httpx==0.24.1
python-jose==3.3.0
python_multipart==0.0.6
jinja2==3.1.2
aiofiles==23.1.0






